Rename crit section HW-specific file
